requirements.txt
streamlit>=1.30.0
requests>=2.31.0
anthropic>=0.18.0
streamlit-agraph>=0.0.45
streamlit-drawable-canvas>=0.9.3
Pillow>=10.0.0
pandas>=2.0.0
reportlab>=4.0.0
